Significantly alter clustering algorithm to resolve issues with Orbit incorrectly causing multiple activations of the same actor in a single cluster. These issues were specifically observed using the orbit-redis-cluster plugin, not the builtin JGroupsClusterPeer.
Old cluster algorithm has nodes joining a cluster by announcing that they are members and the immediately proceeding without waiting for agreement from other nodes. This is bad because two nodes in the “same” cluster might not ‘see’ each other for several seconds. During this time, they will assume that any actor directory entries for the other node are defunct and will replace them, causing the multiple-activation issue above.

The old implementation also had the `Hosting` (aka `NodeCapabilities`) layer implement queries between nodes about which actors are supported where. The new implementation pushes those details down into the `ClusterPeer` implementation, which can then be simply based on a heartbeat-type algorithm.

Now the `ClusterPeer` is responsible for making up-to-date `ClusterView` implementations available to the `Hosting` layer, which uses that `ClusterView` to perform its duties, including checking the `ActorDirectory` entries as appropriate.

This has had significant fallout:

* JGroupsClusterPeer no longer works, because I didn't re-implement the queries from above on top of JGroups.
* ActorClassFinder instances are now no longer permitted to be lazy (because nodes will announce their available set of actors as soon as they start up). `LazyActorClassFinder` is retained only for its role in unit tests.
* Stage.pulse is now called 1/sec rather than 1/10sec. If this frequency becomes a problem then we will need to decouple Stage.pulse from ClusterPeer.pulse.
* Actors will no longer "time out" while "waiting for a server to host the X" actor. If no such nodes are available, the actor will fail its activation immediately.

Also, an unrelated issue with `@OnlyWithActivated` invocations potentially triggering actor activations was found and fixed -- if an Invocation was issued for an actor that _seemed_ to be activated based on the ActorDirectory, but which was not actually activated, then the actor would be activated to serve that Invocation on the receive side. Invocations will now be discarded on both sender and receiver sides if the actor appears to be deactivated.